reasoning_trace_project/requirements.txt
transformers
accelerate
torch
sentencepiece
protobuf
tqdm
pandas
numpy
matplotlib
pyyaml
